Fix it future foster

+++ app/src/main/java/com/foster/vaccination/VaccinationInfo.kt
@@ -5,4 +5,4 @@ import kotlinx.parcelize.Parcelize
 import java.util.*
 
 @Parcelize
-data class VaccinationInfo(val country : String, val timeline : SortedMap<String, Int>): Parcelable
+data class VaccinationInfo(val country : String, val timeline : SortedMap<String, Long>): Parcelable
